Remove obsolete single-file component
