+++ README.md
@@ -1,1 +1,14 @@
-# college-helper
+# College Helper
+
+This project provides an interactive college admission predictor built with React and Vite. It allows students to input their academic profile and get admission predictions for various schools. College statistics are fetched from Google's Gemini API when an API key is provided.
+
+## Development
+
+Install dependencies and start the development server:
+
+```bash
+npm install
+npm run dev
+```
+
+Create a `.env` file with your `GEMINI_API_KEY` to enable live college data fetching.
